Add notes to clarify functionality
it was by chance that I came across this thread while researching for my current project of terraforming our move from AWS cognito to Auth0.

https://community.auth0.com/t/management-api-to-create-a-tenant-and-to-install-and-configure-an-extension/40537

I think it would be really nice to have this information more prominent. I need to create 4 tenants and install some extensions in them. I was expecting to do all of that with the auth0_tenant resource because typically TF can create and destroy most all resource types. I think not being able to create/destroy the resource is an exception to the norm.

BTW - this is my first time proposing changes to something public like this. Hope I didn't bungle it.